feat: Initial version of Home Run (#1)
This PR implements the initial version of Home Run. The original version
lived in multiple Investec Azure DevOps repos.
## How is this put together?

We've used [Create TypeScript App
](https://github.com/JoshuaKGoldberg/create-typescript-app) to scaffold
out a respository. CTA is designed to scaffold TypeScript apps in
GitHub. I should probably mention that I'm a contributor to that
project. It's very useful.

As a consequence, this repo includes GitHub workflows that cover the
following:
- building
- linting
- formatting
- testing
- publishing (using CI/CD and semantic releases)

## Copyright

As per [our
documentation](https://oneinvestec.atlassian.net/wiki/spaces/ICEX/pages/219962846/License+and+Copyright),
our codefiles feature a header that asserts Investec's Copyright:

```ts
/**
 * Copyright (c) Investec
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE.md file in the root directory of this source tree.
 */
```

## Contribution guide

As per [our
documentation](https://oneinvestec.atlassian.net/wiki/spaces/ICEX/pages/220036857/How+to+get+started+contributing),
the repository features a [guide on how to
contribute](https://github.com/investec/home-run/blob/prepare/.github/CONTRIBUTING.md).
This includes a code of conduct and a guide on how to get up and running
with development.

## Still to do

At the time of writing, this is a **private** repository on GitHub. It
will remain so until we feel ready to hit the "Public" button. Before
that we will share with Architecture and Engineering Steercom. The
document that will be shared with Architecture and Engineering Steerco
proposing the OSSing of Home Run can be [found
here](https://investeceu-my.sharepoint.com/:w:/g/personal/john_reilly_investec_com/Eds8CqWM1nRHiibDZ3os2tUBlL3PSdKt4kjVba8coDi6Sw?e=333TAn).

### Scanning

One the repo is public (has been open sourced) then we can start to plug
in workflows such as GitHub Advanced Security that will handle
dependency upgrades etc. We need our repo to be public in order to be
able to use them.

However, with the help of @irfanj01, @rick-roche and @ctaceygreen, we
have tested this repo in Azure DevOps using GitHub Advanced Security for
Azure DevOps and found a clean bill of health:


https://dev.azure.com/investec/sandbox/_build/results?buildId=1164607&view=logs&j=70764e4f-eaf2-5618-fd0d-39e528c90335&t=92265a00-6845-5c35-4128-4c776c06c827

<img width="1139" alt="screenshot of clean bill of health"
src="https://github.com/user-attachments/assets/2223b848-2eef-440f-836a-1c5bdce69f2c">

### Publishing to npm

We have not configured publishing to npm. We will do this with the help
of @marktew once we have approval from A&E SteerCo.

### General tidy

There are references in the
[`DEVELOPMENT.md`](https://github.com/investec/home-run/blob/prepare/.github/DEVELOPMENT.md)
to Investec specific projects. These will replaced with something more
generic prior to OSSing. They are useful now for demo purposes. Possibly
some other bits and pieces.

We will look to do this before merging this PR. We aim not to merge this
PR until A&E SteerCo have approved.

### More tests!

We have automated tests in place, we had none on the original codebase.
We should increase the coverage further and we will.

## The Home Run codebase

The Home Run codebase was taken from an internal repository in
Investec's Azure DevOps

As it has been moved into the new codebase, we have enhanced it to make
use of some nice affordances taken from CTA such as spinners /
formatters etc.

Rather than repeat the docs for Home Run in this PR, it's worth reading
the
[README.md](https://github.com/investec/home-run/blob/prepare/README.md).